requirements.txt
absl-py==0.7.0
astor==0.7.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dlib==19.22.0
editdistance==0.5.2
Flask==1.1.2
gast==0.3.3
glog==0.3.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
kaldiio==2.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
networkx==2.5.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
pyparsing==2.4.7
pyquaternion==0.9.9
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.4.1
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tqdm==4.51.0
typing-extensions==3.10.0.0
urllib3==1.26.2
Werkzeug==1.0.1
win32-setctime==1.0.3
wrapt==1.12.1
zipp==3.4.0
